restructured file architecture
